todo: clean up lots of mathy code todo with render boxes
