+++ src/lib.rs
@@ -108,5 +108,5 @@ pub use user_select::UserSelectNone;
 mod user_select;
 pub mod utils;
 mod widget;
-pub(crate) use render_targets::{ChangedCosmicWidgetSize, CosmicWidgetSize, SourceType};
+pub(crate) use render_targets::{ChangedCosmicWidgetSize, CosmicWidgetSize};
 pub mod render_targets;
